requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
rdkit==2024.9.6
Pillow